+++ app/Http/Controllers/QuizController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Topic;
 use App\Models\Answer;
 use App\Models\UserAnswer;
+use App\Models\Question;
 
 use Illuminate\Http\Request;
 use App\Models\UserTopicResult;
@@ -57,6 +58,31 @@ class QuizController extends Controller
             'user_score' => $user_score,
             'user_score_to_hundred' => $user_score_to_hundred,
         ]);
+
+        $topic_id = $request->topic_id;
+        $questions = Question::where('topic_id', $topic_id)->get();
+
+        foreach ($questions as $question) {
+            $answers[$question->id] = Answer::where('question_id', $question->id)->get();
+        }
+
+        return view(
+            'quiz.answers',
+            [
+                'topic_id' => $topic_id,
+                'questions' => $questions,
+                'answers' => $answers
+            ]
+        );
+    }
+
+    public function result(Request $request)
+    {
+        $topic_completed = UserTopicResult::where('topic_id', $request->topic_id)
+            ->where('user_id', auth()->user()->id)
+            ->first();
+
+        $user_score_to_hundred = $topic_completed->user_score_to_hundred;
 
         return view('quiz.result', ['user_score_to_hundred' => $user_score_to_hundred]);
     }