+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Development
 
+* 2024-11.11 -- 0.6.4 - Practical tasks
+  * Show results after completing test
+
 * 2024-10-26 -- 0.6.3 - Practical tasks
   * Added Practical Tasks - user can send text answer to task and then this answer can be reviewed
   * Bug fixes